requirements.txt
Django==2.0.1
amqp==2.2.1
anyjson==0.3.3
Babel==2.5.0
billiard==3.5.0.3
celery==4.1.0
celery-haystack==0.10
chardet==3.0.4
contextlib2==0.5.5
diff-match-patch==20121119
django-admin-list-filter-dropdown==1.0.1
django-admin-rangefilter==0.3.0
django-appconf==1.0.2
django-celery==3.2.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-daterange-filter==1.3.0
django-import-export==0.7.0
et-xmlfile==1.0.1
flower==0.9.2
jdcal==1.3
kombu==4.1.0
odfpy==1.3.5
openpyxl==2.4.8
psycopg2==2.7.2
pytz==2017.3
PyYAML==3.12
raven==6.1.0
redis==2.10.6
tablib==0.12.1
tornado==4.5.2
unicodecsv==0.14.1
vine==1.1.4
xlrd==1.1.0
xlwt==1.3.0